+++ src/app/locales/en/components/acl-editor.js
@@ -1,3 +1,5 @@
+import aceEditor from './acl-editor/ace-editor';
+
 export default {
   permissionGroups: {
     data: 'Data',
@@ -22,28 +24,8 @@ export default {
     write_attributes: 'Write attributes',
     delete: 'Delete',
   },
-  aceEditor: {
-    aceType: 'Type',
-    aceTypes: {
-      allow: 'Allow',
-      deny: 'Deny',
-    },
-    acePermissionState: {
-      deny: 'denied',
-      allow: 'allowed',
-    },
-    moveDown: 'Move down',
-    moveUp: 'Move up',
-    remove: 'Remove',
-    unknown: 'Unknown',
-    id: 'ID',
-    aceSubjects: {
-      user: 'user',
-      group: 'group',
-    },
-    aceNotAccessible: 'This Access Control Entry refers to a {{subject}} that is no longer a member of this space.',
-  },
   addUserOrGroup: 'Add user or group...',
   noAce: 'No Access Control Entries specified.',
   posixPermissionsWillApply: 'POSIX permissions will be applied.',
+  aceEditor,
 };