+++ src/app/locales/en/components/share-show/pane-opendata.js
@@ -3,6 +3,9 @@ import { MetadataType } from 'oneprovider-gui/models/handle';
 export const handleServiceTextMore =
   'Handle service is used as a mediator to register a persistent identifier for your data collection. Handle services are typically configured by the managers in your organization or a Onezone admin, who can add members (users or groups) and grant them the privileges to publish new records.';
 
+export const noHandleServicesText =
+  'You do not have access to any handle service, which is necessary to register Open Data records. You can either create a new handle service (requires <code>oz_handle_services_create</code> admin privilege) or request access to an existing handle service from the site administrators.';
+
 export default {
   publishWelcome: {
     title: 'Publish as Open Data',
@@ -27,7 +30,6 @@ export default {
   },
   noHandleServices: {
     title: 'No handle services available',
-    content: 'You do not have access to any Open Data handle service, which is necessary to register Open Data records. You can either create a new handle service (requires <code>oz_handle_services_create</code> admin privilege), or contact the site administrators and ask for access to an existing handle service.',
+    content: noHandleServicesText,
   },
-  noMetadata: 'This dataset has no metadata attached.',
 };